added metadata parsing
Parsing is not very sophisticated. It just parses tags from a single IMG table.

It returns these results as a dictionary with the keys being the field names on IMG.

This has not been tested but I intend to test it after merging with master and will
fix any issues that come up them.

Later, there should be improvements to parse metadata from JGI as well.